ct_dashboard_api: extract invitation acceptance logic to api layer
