+++ backend/ct_dashboard_api/src/routes/v1/validate_invitation.ts
@@ -1,15 +1,12 @@
 import { ErrorCodeMap } from "@oko-wallet/oko-api-error-codes";
 import { registry } from "@oko-wallet/oko-api-openapi";
 import { ErrorResponseSchema } from "@oko-wallet/oko-api-openapi/common";
-import { getCTDUserWithCustomerByEmail } from "@oko-wallet/oko-pg-interface/customer_dashboard_users";
-import {
-  getTeamInvitationByToken,
-  updatePendingInvitationStatus,
-} from "@oko-wallet/oko-pg-interface/customer_team_invitations";
 import type { OkoApiResponse } from "@oko-wallet/oko-types/api_response";
 import type { ValidateInvitationRequest } from "@oko-wallet/oko-types/ct_dashboard";
 import type { Request, Response } from "express";
 import { z } from "zod";
+
+import { validateTeamInvitation } from "@oko-wallet-ctd-api/api/team_invitation";
 
 const ValidateInvitationRequestSchema = registry.register(
   "ValidateInvitationRequest",
@@ -77,106 +74,14 @@ export async function validateInvitation(
   req: Request<unknown, unknown, ValidateInvitationRequest>,
   res: Response<OkoApiResponse<unknown>>,
 ) {
-  try {
-    const state = req.app.locals;
-    const { token } = req.body;
+  const state = req.app.locals;
 
-    const invitationRes = await getTeamInvitationByToken(state.db, token);
+  const result = await validateTeamInvitation(state.db, req.body);
 
-    if (!invitationRes.success) {
-      res.status(500).json({
-        success: false,
-        code: "UNKNOWN_ERROR",
-        msg: invitationRes.err,
-      });
-      return;
-    }
-
-    if (invitationRes.data === null) {
-      res.status(ErrorCodeMap.INVITATION_NOT_FOUND).json({
-        success: false,
-        code: "INVITATION_NOT_FOUND",
-        msg: "Invitation not found",
-      });
-      return;
-    }
-
-    const invitation = invitationRes.data;
-
-    if (invitation.status !== "PENDING") {
-      const code =
-        invitation.status === "ACCEPTED"
-          ? "INVITATION_ALREADY_ACCEPTED"
-          : "INVITATION_NOT_PENDING";
-      res.status(ErrorCodeMap[code]).json({
-        success: false,
-        code,
-        msg: `Invitation is ${invitation.status.toLowerCase()}`,
-      });
-      return;
-    }
-
-    if (new Date(invitation.expires_at) < new Date()) {
-      await updatePendingInvitationStatus(
-        state.db,
-        invitation.invitation_id,
-        "EXPIRED",
-      );
-      res.status(ErrorCodeMap.INVITATION_EXPIRED).json({
-        success: false,
-        code: "INVITATION_EXPIRED",
-        msg: "Invitation has expired",
-      });
-      return;
-    }
-
-    // Check if email already belongs to a team
-    const existingUserRes = await getCTDUserWithCustomerByEmail(
-      state.db,
-      invitation.email,
-    );
-    if (existingUserRes.success && existingUserRes.data !== null) {
-      res.status(ErrorCodeMap.DUPLICATE_TEAM_MEMBER).json({
-        success: false,
-        code: "DUPLICATE_TEAM_MEMBER",
-        msg: "This email is already associated with a team",
-      });
-      return;
-    }
-
-    // Verify customer is still active
-    const customerQuery = `
-      SELECT label FROM customers
-      WHERE customer_id = $1 AND status = 'ACTIVE'
-    `;
-    const customerResult = await state.db.query(customerQuery, [
-      invitation.customer_id,
-    ]);
-    if (customerResult.rows.length === 0) {
-      res.status(ErrorCodeMap.INVITATION_NOT_FOUND).json({
-        success: false,
-        code: "INVITATION_NOT_FOUND",
-        msg: "The team no longer exists",
-      });
-      return;
-    }
-    const teamName = customerResult.rows[0].label;
-
-    res.status(200).json({
-      success: true,
-      data: {
-        email: invitation.email,
-        role: invitation.role,
-        team_name: teamName,
-      },
-    });
-    return;
-  } catch (_error) {
-    res.status(500).json({
-      success: false,
-      code: "UNKNOWN_ERROR",
-      msg: "Internal server error",
-    });
+  if (!result.success) {
+    res.status(ErrorCodeMap[result.code] ?? 500).json(result);
     return;
   }
+
+  res.status(200).json(result);
 }
